+++ packages/ide-extension/src/telemetry/telemetry.ts
@@ -5,7 +5,7 @@ import { TelemetryClient } from 'applicationinsights';
 import type { Contracts } from 'applicationinsights';
 import type { IDE, SendTelemetry } from '@sap/guided-answers-extension-types';
 import { v4 as uuidv4 } from 'uuid';
-import { logString } from '../logger/logger';
+import { logString, traceString } from '../logger/logger';
 import packageJson from '../../package.json';
 import type { TelemetryEvent, TelemetryReporter } from '../types';
 import { actionMap } from './action-map';
@@ -101,7 +101,7 @@ export async function trackEvent(event: TelemetryEvent): Promise<void> {
         const name = `${packageJson.name}/${event.name}`;
         const properties = propertyValuesToString({ ...event.properties, ...(reporter.commonProperties ?? {}) });
         reporter.client.trackEvent({ name, properties });
-        logString(`Telemetry event '${event.name}': ${JSON.stringify(properties)}`);
+        traceString(`Telemetry event '${event.name}': ${JSON.stringify(properties)}`);
     } catch (error) {
         logString(`Error sending telemetry event '${event.name}': ${(error as Error).message}`);
     }