+++ packages/ide-extension/test/logger/logger.test.ts
@@ -1,5 +1,5 @@
 import { OutputChannel, window } from 'vscode';
-import { logString } from '../../src/logger/logger';
+import { logString, traceString } from '../../src/logger/logger';
 
 describe('Tests for logging', () => {
     test('Test for logString()', () => {
@@ -14,5 +14,18 @@ describe('Tests for logging', () => {
 
         // Result check
         expect(channelMock.appendLine).toBeCalledWith('LOG_MESSAGE');
+    });
+
+    test('Test for traceString()', () => {
+        // Mock setup
+        const originalLogFn = console.log;
+        console.log = jest.fn();
+
+        // Test execution
+        traceString('TRACE_MESSAGE');
+
+        // Result check
+        expect(console.log).toBeCalledWith('TRACE_MESSAGE');
+        console.log = originalLogFn;
     });
 });