+++ packages/types/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap/guided-answers-extension-types
 
+## 1.25.0
+
+## 1.24.2
+
+## 1.24.1
+
+### Patch Changes
+
+-   34c9138: Fix for duplicate search results
+
+## 1.24.0
+
+### Minor Changes
+
+-   37a50f4: Read HTML enhancements from API
+
 ## 1.23.7
 
 ### Patch Changes